+++ Sources/BraintreeCore/BTAPIClient.swift
@@ -332,6 +332,7 @@ import Foundation
     @_documentation(visibility: private)
     public func sendAnalyticsEvent(
         _ eventName: String,
+        isVaultRequest: Bool? = nil,
         correlationID: String? = nil,
         errorDescription: String? = nil,
         linkType: String? = nil,
@@ -339,6 +340,7 @@ import Foundation
     ) {
         analyticsService?.sendAnalyticsEvent(
             eventName,
+            isVaultRequest: isVaultRequest,
             correlationID: correlationID,
             errorDescription: errorDescription,
             linkType: linkType,