+++ UnitTests/BraintreeCoreTests/Analytics/FakeAnalyticsService.swift
@@ -7,6 +7,7 @@ class FakeAnalyticsService: BTAnalyticsService {
 
     override func sendAnalyticsEvent(
         _ eventName: String,
+        isVaultRequest: Bool? = nil,
         correlationID: String? = nil,
         endpoint: String? = nil,
         endTime: Int? = nil,